+++ R/worldplot.R
@@ -12,9 +12,12 @@
 #' @param crs Coordinate reference system (EPSG). By default the value is 4326, which corresponds to EPSG::4326 (WGS84)
 #' @param title Title of the plot. Default is no title.
 #' @param legendTitle Title of the legend. Default is the name of the filling variable.
+#' @param legend.position Position of the legend. If set to "none", no legend is displayed
 #' @param annote Do you want to plot country labels (ISO 3166-1 alpha-2 code) on the map? Default is set to \code{FALSE}.
 #' @param div Parameter for modifying the elements dimensions in the map. Usually, it does not need to be modified. Default value is 1.
 #' @param palette_option Character string indicating the palette to be used. Available options range between "A" and "H".
+#' @param label.color Color of the labels if annote = TRUE. Default is white
+#' @param label.size Size of the labels if annote = TRUE
 #' @param na_colour The colour to be used for countries with missing information. Default is grey80
 #' @param transform_limits Only if crs is specified and different from 4326. If TRUE (the default) the program expects to receive values of longitude and latitude as in EPSG 4326,
 #'                          (i.e., within -180, +180 for longitude and within -90, +90 for latitude) and automatically updates to the new crs.
@@ -36,8 +39,8 @@
 worldplot <- function(data,
                       ColName, CountryName, CountryNameType = "isoa2", rangeVal,
                       longitude = c(-180, 180) ,latitude = c(-90, 90), crs = 4326,
-                      title = "", legendTitle = as.character(ColName),
-                      annote = FALSE, div = 1, palette_option = "D",
+                      title = "", legendTitle = as.character(ColName), legend.position = "right",
+                      annote = FALSE, div = 1, palette_option = "D", label.color = "white", label.size = 2,
                       na_colour = "grey80", transform_limits = TRUE) {
 
   world <- ne_countries(scale = 50, continent = NULL, returnclass = "sf")
@@ -79,17 +82,28 @@ worldplot <- function(data,
     rangeVal = c(range(map_df$MapFiller, na.rm = TRUE))
   }
   
+  if (crs != 4326 & transform_limits == TRUE) {
+      #Correct longitude and latitude values
+      lim1 <- data.frame(X = longitude, Y = latitude)
+      lim2 <- st_as_sf(lim1, coords=c("X","Y"), crs="EPSG:4326" )
+      lim3 <- st_transform(lim2, crs = st_crs(crs))
+      
+      longitude <- c(st_bbox(lim3)$xmin, st_bbox(lim3)$xmax)
+      latitude <-  c(st_bbox(lim3)$ymin, st_bbox(lim3)$ymax)
+      ##
+    }
 
   wplot <- ggplot(data= map_df) +
     geom_sf(color= 'black', aes(fill= MapFiller)) +
     theme(legend.key.size = unit(1/div, 'lines'),
           legend.text = element_text(size= 8/div),
           legend.title = element_text(size= 8/div),
+          legend.position = legend.position,
           plot.title = element_text(size=8/div),
           panel.grid = element_blank(),
           panel.background = element_rect(fill = 'grey95'))+
     labs(fill= legendTitle)+
-    coord_sf(xlim= longitude, ylim= latitude, expand= FALSE, label_axes = 'SW') +
+    coord_sf(xlim= longitude, ylim= latitude, expand= FALSE, label_axes = 'SW', crs = st_crs(crs)) +
     xlab('') + ylab('')+
     ggtitle(title)
   
@@ -105,53 +119,14 @@ worldplot <- function(data,
   }
   
 
-  if (crs != 4326) {
-    
-    if (transform_limits == TRUE) {
-    #Correct longitude and latitude values
-    lim1 <- data.frame(X = longitude, Y = latitude)
-    lim2 <- st_as_sf(lim1, coords=c("X","Y"), crs="EPSG:4326" )
-    lim3 <- st_transform(lim2, crs = st_crs(crs))
-    
-    longitude <- c(st_bbox(lim3)$xmin, st_bbox(lim3)$xmax)
-    latitude <-  c(st_bbox(lim3)$ymin, st_bbox(lim3)$ymax)
-    ##
-    }
-    
-    wplot <- wplot +
-      coord_sf(xlim= longitude, ylim= latitude, expand= FALSE, label_axes = 'SW',
-               crs = st_crs(crs))
-  }
-
   if (annote == TRUE) {
 
-    world_points <- geometries_data(exclude.iso.na = T,
-                                    countries.list = simdata$iso_a2[!is.na(simdata$MapFiller)])
-
-    if (crs != 4326) {
-
-      d <- data.frame(iso_a2 = world_points$iso_a2,
-                      X = world_points$X,
-                      Y =world_points$Y)
-
-      d2 <- st_as_sf(d, coords=c("X","Y"), crs="EPSG:4326" )
-
-      d3 <- st_transform(d2, crs = st_crs(crs))
-
-      d4 <- data.frame(iso_a2 = d3$iso_a2,
-                       X = rep(NA, nrow(d3)),
-                       Y = rep(NA, nrow(d3)))
-
-      for (i in 1: nrow(d3)) {
-        d4[i,c("X","Y")] <- d3$geometry[[i]]
-      }
-
-      world_points <- d4
-
-    }
+    world_points <- countrycoord_data(countries.list = simdata$iso_a2[!is.na(simdata$MapFiller)],
+                                      crs = crs, UK_as_GB = TRUE, exclude.iso.na = TRUE)
 
     wplot <- wplot +
-      with_shadow(geom_text(data= world_points, aes(x=X, y=Y,label= iso_a2), size= 2/div, color= 'white', fontface= 'bold'),
+      with_shadow(geom_text(data= world_points, aes(x=X, y=Y,label= iso_a2), size= label.size/div, 
+                            color= label.color, fontface= 'bold'),
                   x_offset = 2, y_offset = 2, sigma = 1)
   }